+++ packages/mint-components/src/components/sqm-password-field/usePasswordField.tsx
@@ -32,7 +32,7 @@ export function usePasswordField(
 
   return {
     states: {
-      enableValidation: props.enableValidation,
+      enableValidation: !props.disableValidation,
       dynamicValidation,
       registrationFormState,
       validationErrors: validationState?.validationErrors,