+++ packages/mint-components/src/components/sqm-portal-register/usePortalRegister.tsx
@@ -121,6 +121,7 @@ export function usePortalRegister(props: PortalRegister) {
       confirmPassword: props.confirmPassword,
       hideInputs: props.hideInputs,
       loginPath: props.loginPath,
+      disablePasswordValidation: !props.disablePasswordValidation,
     },
     callbacks: {
       submit,